// starters/graphql-serverless-contentful/src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Technology: GET */
  technology?: Maybe<Array<Maybe<Technology>>>;
};


export type QueryTechnologyArgs = {
  id?: InputMaybe<Scalars['ID']>;
};

export type Technology = {
  __typename?: 'Technology';
  description?: Maybe<Scalars['String']>;
  displayName: Scalars['String'];
  id: Scalars['ID'];
  url?: Maybe<Scalars['String']>;
};